requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
bcrypt==4.1.2
python-dotenv==1.0.1
google-genai==0.3.0